week-5 create blueprint
